Server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
bcrypt==4.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
dacite==1.8.1
databases==0.9.0
debugpy==1.8.9
decorator==5.1.1
distlib==0.3.8
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
Events==0.5
executing==2.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.15.4
firebase-admin==6.7.0
fonttools==4.53.1
fsspec==2024.12.0
gevent==24.11.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-genai==1.7.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
haven==2.0.4
htmlmin==0.1.12
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.6
ImageHash==4.3.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mpmath==1.3.0
msgpack==1.1.0
multimethod==1.12
nest-asyncio==1.6.0
netkit==3.1.12
networkx==3.4.2
numba==0.60.0
numpy==2.0.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.3.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.2
pyreadstat==1.2.8
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.1
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.4.0
setproctitle==1.3.4
setuptools==78.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==3.1.0
typeguard==4.4.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.26.3
visions==0.7.6
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
ydata-profiling==4.12.0
zope.event==5.0
zope.interface==7.2
